use trimalloc() for arrays that later get trifree()d #1
Crashes on Win64 pointed to this mismatch where pointlists and
other arrays allocated with Perl macros were later free()d using
trifree() (which is just free()). So now we allocate with trimalloc()
(just malloc really), and we've made sure "void" gets used where we
need void pointers, instead of "int", which isn't always the
right size to hold a pointer.